move test file to scenario folder
